backend/requirement.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pymongo
motor
langchain
langgraph
openai
httpx